2006-08-16  Al Riddoch  <alriddoch@zepler.org>
	* wfmath/intstring_test.cpp: Switch to using atol instead of
	  atoi to ensure longs work on platforms where long is bigger
	  than int.

+++ wfmath/intstring_test.cpp
@@ -40,7 +40,7 @@ static void TestConvert()
     // random number, so we don't care. Large unsigned ints will
     // provide us negative numbers for testing.
     long val2 = (long) val;
-    assert(atoi(IntToString(val2).c_str()) == val2);
+    assert(atol(IntToString(val2).c_str()) == val2);
   }
 }
